Renamed textAnalysisAPI.properties to Aylien.properties, caused changes in .gitignore, SummarizeService.java, and RetreiveProperties.java, added a lot to Filter.java
