+++ scitweets/src/main/java/aylienProperties/RetrieveProperties.java
@@ -1,4 +1,4 @@
-package textAnalysisAPIProperties;
+package aylienProperties;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -8,14 +8,15 @@ import java.util.Properties;
 
 public class RetrieveProperties {
 	public static String getKey() throws FileNotFoundException, IOException {
-		File configFile = new File("textAnalysisAPI.properties");
+		File configFile = new File("Aylien.properties");
 		FileReader reader = new FileReader(configFile);
 		Properties prop = new Properties();
 		prop.load(reader);
 		return prop.getProperty("KEY");
 	}
+
 	public static String getAppID() throws FileNotFoundException, IOException {
-		File configFile = new File("textAnalysisAPI.properties");
+		File configFile = new File("Aylien.properties");
 		FileReader reader = new FileReader(configFile);
 		Properties prop = new Properties();
 		prop.load(reader);
